pyproject.toml
[project]
name = "internia"
version = "0.1.0"
description = ""
authors = [
    {name = "Valdeir Sampaio Sapará Filho",email = "saparavaldeir@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-vite (>=3.1.0,<4.0.0)",
    "inertia-django (>=1.2.0,<2.0.0)"
]

[tool.poetry]
package-mode = false
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
